test: drop unnecessary f64 casts (clippy 1.95 / CI fix)
CI's stable toolchain advanced to rust 1.95.0, whose clippy flags
`Seconds::value() as f64` (f64 -> f64) in the B6 PSD test as
unnecessary_cast under -D warnings. Pre-existing; first surfaced now
because earlier pushes failed the fmt gate before reaching clippy.
Verified with the CI command on a local 1.95.0 toolchain.

+++ crates/engine/tests/clinical_perturbation.rs
@@ -262,8 +262,8 @@ fn psd_tracks_period_jitter() {
     // Period standard deviation (PSD): the jittered signal has a wider
     // spread of periods than the clean one, so a larger PSD. Both are
     // small and positive.
-    let clean = measure("clean_120hz").period_std_s.value() as f64;
-    let jittered = measure("jitter_150hz").period_std_s.value() as f64;
+    let clean = measure("clean_120hz").period_std_s.value();
+    let jittered = measure("jitter_150hz").period_std_s.value();
     assert!(
         clean >= 0.0 && jittered > 0.0,
         "psd clean {clean}, jit {jittered}"
